Warning fixed.

+++ Pod/Classes/MagnetPickerViewController.m
@@ -125,13 +125,11 @@
     [self.pickerView reloadAllComponents];
 }
 
--(NSInteger)pickerView:(UIPickerView *)pickerView numberOfRowsInComponent:(NSInteger)component
-{
+-(NSInteger)pickerView:(UIPickerView *)pickerView numberOfRowsInComponent:(NSInteger)component {
     return [self.filteredOptions count];
 }
 
--(int)numberOfComponentsInPickerView:(UIPickerView *)pickerView
-{
+-(NSInteger)numberOfComponentsInPickerView:(UIPickerView *)pickerView {
     return 1;
 }
